Aligned header content
First I switched the display to 'flex' for the body element. It made my content into columns running from left to right. I changed the 'flex-direction' from row to column and my content stacked vertically like I wanted. Then I did the same for the header class. My header links were stacked as well so I changed the display of 'ul' to flex. Since I wanted my header logo to be on the left and the tabs to the right, I aligned everything to the right (flex-end). Then I changed the logo's right margin to 'auto', making it go to the left. I changed the 'align-text' to 'center', added a gap between the header links, and increased the margin size (left & right).