Align hero section
I aligned the text within the hero section to the left of the webpage. I created a box for the image placeholder and moved it on the right side of the page. I made the 'sign up' button bigger. I also increased the font size of the text and made it bolder.