Edit Sign Up Box
I made the text white and the background color blue with rounded corners. I also added a light blue 'Sign Up' button like the one in the hero section. I aligned the text to the left within the box and the button to the right. I added some padding at the bottom of the container class, between the sign up box and the footer.